requirements.txt
#libraries need to be installed 
streamlit 
google-generativeai 
python-dotenv 
PyPDF2 
Pillow
plotly